requirements.txt
# Core runtime
pandas>=2.1
numpy>=1.26
pyyaml>=6.0
pydantic>=2.7
scikit-learn>=1.4
xgboost>=2.0
shap>=0.45
matplotlib>=3.8

# GIS
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
fiona>=1.9
rasterio>=1.3
rasterstats>=0.19
folium>=0.16
branca>=0.7

# NetCDF reading (for CHIRPS/ERA5 style climate inputs)
netCDF4>=1.7

# CLI
typer>=0.12
rich>=13.7

# Output formats
pyarrow>=16.0
